Rename projects Fixes #8
