+++ Engine/Graphics/IResourceRepository.cs
@@ -19,7 +19,7 @@
  CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-namespace Graphics.Interfaces;
+namespace Engine.Graphics;
 
 public interface IResourceRepository
 {
