+++ Physics/Physics.csproj
@@ -1,4 +1,4 @@
-/*
+<!--
  Copyright (c) 2022 Loïc Morvan
 
  Permission is hereby granted, free of charge, to any person obtaining a copy of
@@ -17,25 +17,14 @@
  COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
  IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
  CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
+ -->
 
-using Pinball.Impl;
+<Project Sdk="Microsoft.NET.Sdk">
 
-namespace Pinball.Tests;
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
 
-public class CollisionResolverTests
-{
-    [Fact]
-    public void Test()
-    {
-        var ball = new Ball(new(0, -1), 0, 0.25m);
-        var collision = new Collision(0.5m, new(0, -0.75m), new(0, 1), 1);
-        var sut = new CollisionResolver();
-
-        var result = sut.ResolveCollision(ball, 1, collision);
-
-        Assert.NotNull(result);
-        Assert.Equal(new(0, 1), result.s);
-        Assert.Equal(0, result.X);
-    }
-}
+</Project>
