+++ Pinball/Program.cs
@@ -21,12 +21,11 @@
 
 using System.Reflection;
 using Engine;
-using Graphics.Implementations;
-using Pinball;
-using Pinball.Impl;
-using Pinball.Interfaces;
+using Engine.Graphics;
+using Physics;
+using Physics.Colliders;
 
-var board = new Board(0.01m, new CollisionResolver());
+var board = new Physics.Board(0.01m, new CollisionResolver());
 board.Ball = new(new(0.1m, 0), new(0, 10), 0.25m);
 board.Colliders = new ICollider[]
 {
@@ -38,7 +37,7 @@ board.Colliders = new ICollider[]
 
 var window = new Window(new AssemblyResourceRepository(Assembly.GetExecutingAssembly()));
 window.Room.Camera = new(-10, 0, 20, 10);
-window.Room.GameObjects.Add(new Pinball.Exe.Board(board));
-window.Room.GameObjects.Add(new GameObject() { Sprite = new("Pinball.Exe.Resources.Sample.png", 10, 10, 5m, 5m) });
-window.Room.GameObjects.Add(new Pinball.Exe.Ball(board) { Sprite = new("Pinball.Exe.Resources.Sample.png", 0.5m, 0.5m, 0.25m, 0.25m) });
+window.Room.GameObjects.Add(new Pinball.Board(board));
+window.Room.GameObjects.Add(new GameObject() { Sprite = new("Pinball.Resources.Sample.png", 10, 10, 5m, 5m) });
+window.Room.GameObjects.Add(new Pinball.Ball(board) { Sprite = new("Pinball.Resources.Sample.png", 0.5m, 0.5m, 0.25m, 0.25m) });
 window.Run();
