--- pinball-physics.sln
@@ -1,40 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30114.105
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Pinball", "Pinball\Pinball.csproj", "{3F9C6D47-BAB4-4EF0-9C74-C30E665DCD41}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Pinball.Tests", "Pinball.Tests\Pinball.Tests.csproj", "{89296A6C-D43A-46CC-9B66-C029C5E27453}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Pinball.Exe", "Pinball.Exe\Pinball.Exe.csproj", "{86A00BE9-710A-4D00-8BB0-37F1FEA8D36B}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Engine", "Engine\Engine.csproj", "{74864A13-E6C5-46F5-8E2F-F8FE7B065A8D}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{3F9C6D47-BAB4-4EF0-9C74-C30E665DCD41}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3F9C6D47-BAB4-4EF0-9C74-C30E665DCD41}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3F9C6D47-BAB4-4EF0-9C74-C30E665DCD41}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3F9C6D47-BAB4-4EF0-9C74-C30E665DCD41}.Release|Any CPU.Build.0 = Release|Any CPU
-		{89296A6C-D43A-46CC-9B66-C029C5E27453}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{89296A6C-D43A-46CC-9B66-C029C5E27453}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{89296A6C-D43A-46CC-9B66-C029C5E27453}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{89296A6C-D43A-46CC-9B66-C029C5E27453}.Release|Any CPU.Build.0 = Release|Any CPU
-		{86A00BE9-710A-4D00-8BB0-37F1FEA8D36B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{86A00BE9-710A-4D00-8BB0-37F1FEA8D36B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{86A00BE9-710A-4D00-8BB0-37F1FEA8D36B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{86A00BE9-710A-4D00-8BB0-37F1FEA8D36B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{74864A13-E6C5-46F5-8E2F-F8FE7B065A8D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{74864A13-E6C5-46F5-8E2F-F8FE7B065A8D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{74864A13-E6C5-46F5-8E2F-F8FE7B065A8D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{74864A13-E6C5-46F5-8E2F-F8FE7B065A8D}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
