+++ pinball.sln
@@ -1,0 +1,40 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30114.105
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Engine", "Engine\Engine.csproj", "{6F193E39-33F6-4986-B46F-10875D22D1C8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Physics", "Physics\Physics.csproj", "{68DCFCD6-8849-4528-9139-D5BB9CE4E732}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Physics.Tests", "Physics.Tests\Physics.Tests.csproj", "{A52933DE-A012-4088-BB65-32B7A59643D8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Pinball", "Pinball\Pinball.csproj", "{9270E686-868E-4913-81A7-8155AC29BBBD}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{6F193E39-33F6-4986-B46F-10875D22D1C8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6F193E39-33F6-4986-B46F-10875D22D1C8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6F193E39-33F6-4986-B46F-10875D22D1C8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6F193E39-33F6-4986-B46F-10875D22D1C8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{68DCFCD6-8849-4528-9139-D5BB9CE4E732}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{68DCFCD6-8849-4528-9139-D5BB9CE4E732}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{68DCFCD6-8849-4528-9139-D5BB9CE4E732}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{68DCFCD6-8849-4528-9139-D5BB9CE4E732}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A52933DE-A012-4088-BB65-32B7A59643D8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A52933DE-A012-4088-BB65-32B7A59643D8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A52933DE-A012-4088-BB65-32B7A59643D8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A52933DE-A012-4088-BB65-32B7A59643D8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9270E686-868E-4913-81A7-8155AC29BBBD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9270E686-868E-4913-81A7-8155AC29BBBD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9270E686-868E-4913-81A7-8155AC29BBBD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9270E686-868E-4913-81A7-8155AC29BBBD}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
